+++ server/seeds/02_app.js
@@ -1,19 +1,16 @@
 const { AppStatus } = require('../src/enums')
-
+const { flatten } = require('../src/utils')
+const users = require('././mock/users')
 const apps = require('./mock/apps')
 const appVersions = require('./mock/appversions')
 const appVersionsLocalised = require('./mock/appversions_localized')
-
-const { flatten } = require('../src/utils')
-
-const users = require('././mock/users')
 
 const statuses = [
     {
         id: 'e900f977-9b5d-496d-9576-cc704359bde1',
         app_id: apps[0].id,
         status: AppStatus.APPROVED,
-        created_by_user_id: users[0].id,
+        created_by_user_id: users[1].id,
     },
     {
         id: '4cb8e2cc-5e92-4b7d-90c4-4ef96e1ea337',
@@ -53,7 +50,7 @@ exports.seed = async knex => {
     await knex('app').del()
     await knex('app').insert(apps)
 
-    console.log('Seeding app statuses:'.statuses)
+    // console.log('Seeding app statuses:', statuses)
 
     await knex('app_status').del()
     await knex('app_status').insert(statuses)
@@ -63,9 +60,9 @@ exports.seed = async knex => {
 
     console.log('Seeding appversions')
 
-    console.log(appVersions)
+    // console.log(appVersions)
     await knex('app_version').insert(flatten(appVersions))
 
-    console.log(appVersionsLocalised)
+    // console.log(appVersionsLocalised)
     await knex('app_version_localised').insert(flatten(appVersionsLocalised))
 }