+++ server/src/routes/v2/apikey.js
@@ -3,17 +3,46 @@ const { ApiKey } = require('../../services/')
 
 module.exports = [
     {
+<<<<<<< HEAD
         method: 'POST',
         path: '/v2/key',
         config: {
             auth: 'token', // you cannot generate a new token using API-key
+=======
+        method: 'GET',
+        path: '/v2/key',
+        config: {
+            auth: 'token',
+>>>>>>> next
             tags: ['api', 'v2'],
         },
         handler: async (request, h) => {
             const { db } = h.context
             const { id: userId } = await getCurrentUserFromRequest(request, db)
 
+<<<<<<< HEAD
             const apiKey = ApiKey.createApiKeyForUser(userId, db)
+=======
+            const apiKey = await ApiKey.getApiKeyByUserId(userId, db)
+
+            return {
+                hasApiKey: !!apiKey,
+                createdAt: apiKey && apiKey.created_at,
+            }
+        },
+    },
+    {
+        method: 'POST',
+        path: '/v2/key',
+        config: {
+            auth: 'token', // you cannot generate a new key using API-key
+            tags: ['api', 'v2'],
+        },
+        handler: async (request, h) => {
+            const { db } = h.context
+            const { id: userId } = await getCurrentUserFromRequest(request, db)
+            const apiKey = await ApiKey.createApiKeyForUser(userId, db)
+>>>>>>> next
             return {
                 apiKey,
             }
@@ -32,7 +61,11 @@ module.exports = [
             const { db } = h.context
             const { id: userId } = await getCurrentUserFromRequest(request, db)
 
+<<<<<<< HEAD
             ApiKey.deleteApiKeyForUser(userId, db)
+=======
+            await ApiKey.deleteApiKeyForUser(userId, db)
+>>>>>>> next
 
             return h.response('API key revoked').code(200)
         },