+++ server/src/routes/v2/apps.js
@@ -1,8 +1,21 @@
+<<<<<<< HEAD
 const { AppStatus } = require('../../enums')
+=======
+const Boom = require('@hapi/boom')
+>>>>>>> next
 const { getApps } = require('../../data')
-const { convertAppsToApiV1Format } = require('../v1/apps/formatting')
-const { filterAppsBySpecificDhis2Version } = require('../../utils/filters')
+const { AppStatus } = require('../../enums')
+const CreateAppModel = require('../../models/v2/in/CreateAppModel')
+const {
+    canCreateApp,
+    getCurrentUserFromRequest,
+    currentUserIsManager,
+} = require('../../security')
+const App = require('../../services/app')
+const Organisation = require('../../services/organisation')
 const Joi = require('../../utils/CustomJoi')
+const { filterAppsBySpecificDhis2Version } = require('../../utils/filters')
+const { convertAppsToApiV1Format } = require('../v1/apps/formatting')
 
 const CHANNELS = ['stable', 'development', 'canary']
 const APPTYPES = ['APP', 'DASHBOARD_WIDGET', 'TRACKER_DASHBOARD_WIDGET']
@@ -21,6 +34,12 @@ module.exports = [
                     )
                         .description('Filter by channel')
                         .default(['stable']),
+<<<<<<< HEAD
+=======
+                    core: Joi.filter(Joi.boolean()).description(
+                        'Filter by core app'
+                    ),
+>>>>>>> next
                     types: Joi.filter(
                         Joi.stringArray().items(Joi.valid(...APPTYPES))
                     )
@@ -38,9 +57,16 @@ module.exports = [
             },
         },
         handler: async (request, h) => {
+<<<<<<< HEAD
             const channels = request.plugins.queryFilter.getFilter('channels')
                 .value
             const types = request.plugins.queryFilter.getFilter('types').value
+=======
+            const queryFilter = request.plugins.queryFilter
+            const channels = queryFilter.getFilter('channels').value
+            const types = queryFilter.getFilter('types').value
+            const coreAppFilter = queryFilter.getFilter('core')
+>>>>>>> next
 
             const apps = await getApps(
                 {
@@ -49,6 +75,7 @@ module.exports = [
                     channels,
                     types,
                     query: request.query.query,
+                    coreApp: coreAppFilter && coreAppFilter.value,
                 },
                 h.context.db
             )
@@ -62,6 +89,87 @@ module.exports = [
                 result,
                 total: result.length,
             })
+        },
+    },
+    {
+        method: 'POST',
+        path: '/v2/apps',
+        config: {
+            auth: 'token',
+            tags: ['api', 'v2'],
+            payload: {
+                maxBytes: 20 * 1024 * 1024, //20MB
+                allow: 'multipart/form-data',
+                parse: true,
+                output: 'stream',
+                multipart: true,
+            },
+            validate: {
+                payload: CreateAppModel.payloadSchema,
+            },
+            plugins: {
+                'hapi-swagger': {
+                    payloadType: 'form',
+                },
+            },
+        },
+        handler: async (request, h) => {
+            if (!canCreateApp(request, h)) {
+                throw Boom.unauthorized()
+            }
+
+            const { db } = h.context
+            const { id: currentUserId } = await getCurrentUserFromRequest(
+                request,
+                db
+            )
+            const isManager = currentUserIsManager(request)
+
+            const { payload } = request
+            const appJsonPayload = JSON.parse(payload.app)
+            const appJsonValidationResult = CreateAppModel.def.validate(
+                appJsonPayload
+            )
+
+            if (appJsonValidationResult.error) {
+                throw Boom.badRequest(appJsonValidationResult.error)
+            }
+
+            const { organisationId } = appJsonPayload.developer
+            const organisation = await Organisation.findOne(
+                organisationId,
+                false,
+                db
+            )
+            if (!organisation) {
+                throw Boom.badRequest('Unknown organisation')
+            }
+
+            const isMember = await Organisation.hasUser(
+                organisationId,
+                currentUserId,
+                db
+            )
+            if (!isMember && !isManager) {
+                throw Boom.unauthorized(
+                    `You don't have permission to upload apps to that organisation`
+                )
+            }
+
+            const { appType } = appJsonPayload
+            const app = await db.transaction(trx =>
+                App.create(
+                    {
+                        userId: currentUserId,
+                        organisationId,
+                        appType,
+                        status: AppStatus.PENDING,
+                    },
+                    trx
+                )
+            )
+
+            return h.response(app).created(`/v2/apps/${app.id}`)
         },
     },
 ]