+++ server/src/routes/v2/index.js
@@ -5,4 +5,5 @@ module.exports = [
     require('./me.js'),
     require('./apikey.js'),
     require('./appVersions'),
+    require('./version.js')
 ]